# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Bejelentkezett felhasznalok listazasa",
                "description": "Bejelentkezett felhasznalok listazasa",
                "operationId": "fe9992e1e7e6e2d260628b2dbd89ef72",
                "responses": {
                    "200": {
                        "description": "sikeres művelet"
                    },
                    "401": {
                        "description": "Nincs bejelentkezve a felhasználó vagy érvénytelen token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/{felhasznaloid}": {
            "patch": {
                "tags": [
                    "user"
                ],
                "summary": "Felhasználó frissítése",
                "description": "Ez csak bejelentkezett felhasználó által tehető meg.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "felhasznaloid",
                        "in": "path",
                        "description": "Frissítendő felhasználó azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Frissített felhasználó objektum",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "nev": "Kovács Balázs",
                                    "email": "kovibali@gmail.com",
                                    "jelszo": "kovibali",
                                    "jelszo_megerositese": "securepassword",
                                    "role": "admin",
                                    "telefonszam": "123456789",
                                    "szemelyi_szam": "12345678",
                                    "szuletesi_datum": "1990-01-01",
                                    "ceg": true,
                                    "cegnev": "Acme Inc.",
                                    "ceg_tipus": "Kft.",
                                    "ado_szam": "123456789",
                                    "bankszamlaszam": "12345678901234567890123456",
                                    "orszag": "Magyarország",
                                    "iranyitoszam": "1234",
                                    "varos": "Budapest",
                                    "utca": "Kossuth utca",
                                    "utca_jellege": "utca",
                                    "hazszam": "10",
                                    "epulet": "A",
                                    "lepcsohaz": "1",
                                    "emelet": "2",
                                    "ajto": "3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Érvénytelen felhasználó adatok"
                    },
                    "404": {
                        "description": "Felhasználó nem található"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/{felhasznaloid}": {
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Felhasználó törlése",
                "description": "Ez csak bejelentkezett felhasználó által tehető meg.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "felhasznaloid",
                        "in": "path",
                        "description": "Törlendő felhasználó azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Érvénytelen felhasználó azonosító"
                    },
                    "404": {
                        "description": "Felhasználó nem található"
                    },
                    "204": {
                        "description": "Sikeres törlés"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}